requirements_gpu.txt
tensorflow-gpu
tensorflow-datasets
keras
git+https://github.com/google-research/nasbench@refs/pull/29/merge
scipy==1.9.1